test: fix update user profile test and return avatar_file from controllers

+++ src/modules/accounts/useCases/updateUserAvatar/UpdateUserAvatarUseCase.ts
@@ -18,9 +18,7 @@ class UpdateUserAvatarUseCase {
   async execute({ user_id, avatar_file }: IRequest): Promise<void> {
     const user = await this.usersRepository.findById(user_id);
 
-    if (user.avatar) {
-      await fileMethods.deleteFile(`./tmp/avatar/${user.avatar}`);
-    }
+    if (user.avatar) await fileMethods.deleteFile(`./tmp/avatar/${user.avatar}`);
 
     user.avatar = avatar_file;
